big update - WOP

+++ wx_gui/databaseTools.py
@@ -1,6 +1,6 @@
 import sqlite3
 from tool import Tool
-
+from gui.menus_inter import MenusInter
 
 def sqlConn():
     # Connect to the SQLite database
@@ -10,8 +10,10 @@ def sqlConn():
 
 
 
-def load_tools_from_database(toolType):
-        
+def load_tools_from_database(toolType, lang):
+   
+        #get the menu text from the dictionary
+        menu = MenusInter(lang)
         #TODO return tools list by toolType
         
         cursor = sqlConn()
@@ -22,7 +24,6 @@ def load_tools_from_database(toolType):
             else:
                 cursor.execute("SELECT * FROM tools WHERE toolType = ? ORDER by D1 ASC", (toolType,))
             tools = cursor.fetchall()
-            #print("Tools readed from database: ", len(tools))
             # add tools to list
             tools_list = []
             existent_tooltypes = []
@@ -32,47 +33,39 @@ def load_tools_from_database(toolType):
                 if tool.toolType not in existent_tooltypes:
                     existent_tooltypes.append(tool.toolType)
                 tools_list.append(tool)
-                #print("tool added: ", tool.name)
                         
-            #tools = reversed(tools) #reverse list to get last added tool first
-                
-            '''
-            #add " " (empty) to dropboxs to clear selection
+            print(f"{menu.get_menu('tools_readed_db').capitalize()} :: {len(tools_list)} :: {len(existent_tooltypes)}")
             
-            self.D1_cb.Append(str(" "))
-            self.L1_cb.Append(str(" "))
-            self.L2_cb.Append(str(" "))
-            self.Z_cb.Append(str(" "))
-            '''
-            print("load_tools_from_database :: ", len(tools_list), " :: ", len(existent_tooltypes))
             return tools_list, existent_tooltypes
         
         except Exception as e:
             #create_db()
-            print("Error: ", e)
+            print("Error loading tools from db: ", e)
             return [], []
 
 def deleteTool(tool):
-    #connect to the database
-    conn = sqlite3.connect('tool_manager.db')
-    #create a cursor
-    cursor = conn.cursor()
-    #delete a record
-    tmp = "DELETE from tools WHERE id='" + str(tool.id) + "'"
-    cursor.execute(tmp)
-    #commit changes
-    conn.commit()
-    #close connection
-    conn.close()
-    print(f":: deleted from database :: {tool.name} :: {tool.toolType}")
+    if tool: 
+        #connect to the database
+        conn = sqlite3.connect('tool_manager.db')
+        #create a cursor
+        cursor = conn.cursor()
+        #delete a record
+        tmp = "DELETE from tools WHERE id='" + str(tool.id) + "'"
+        cursor.execute(tmp)
+        #commit changes
+        conn.commit()
+        #close connection
+        conn.close()
+        print(f":: deleted from database :: {tool.name} :: {tool.toolType}")
 
 
-def delete_selected_item(self, index):
+def delete_selected_item(self, tool):
+    self.panel.olvSimple.RemoveObject(tool)
     
     #delete tool from database
-    deleteTool(self.toolData.full_tools_list[index])
-    #delete tool from list
-    del self.toolData.full_tools_list[index]
+    deleteTool(tool)
+    #delete tool from ovlsimple
+
